Rename module to network_app.py
